requirements.txt
django>=1.4
MySQL-python
python-tastypie
